requirements.txt
# ─── Core Web Framework ─────────────────────────────────────────────────────
Flask==3.1.0
gunicorn==21.2.0
Werkzeug==3.0.1
Jinja2==3.1.3

# ─── Telegram Bot ────────────────────────────────────────────────────────────
python-telegram-bot==21.5

# ─── Firebase Admin SDK ─────────────────────────────────────────────────────
firebase-admin==6.5.0

# ─── Scheduler ───────────────────────────────────────────────────────────────
APScheduler==3.10.4

# ─── Utilidades ──────────────────────────────────────────────────────────────
requests==2.32.2
httpx==0.27.0
